requirements.txt
streamlit==1.14.1
pandas==1.3.4
mysql-connector==2.2.9